mcp_server/requirements.txt
# Core MCP dependencies
mcp>=0.9.0

# Database
psycopg2-binary>=2.9.9

# HTTP and Auth
requests>=2.31.0
authlib>=1.3.0

# JWT handling
PyJWT>=2.8.0

# Environment management
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.7.0
